# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TickerPro",
        "description": "Backend endpoints docs",
        "contact": {
            "email": "support@tickerpro.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://backendbootcamp.proshore.eu/api",
            "description": "Proshore"
        }
    ],
    "paths": {
        "/admin/delete-user/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete User",
                "description": "A user can be deleted by passing their user id.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id of the user whom the admin wants to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid user id"
                    }
                }
            }
        },
        "/admin/view-user/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete User",
                "description": "A user can be deleted by passing their user id.",
                "operationId": "viewUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id of the user whom the admin wants to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid user id"
                    }
                }
            }
        },
        "/admin/assign-user/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete User",
                "description": "A user can be deleted by passing their user id.",
                "operationId": "assignRoles",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id of the user whom the admin wants to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid user id"
                    }
                }
            }
        },
        "/admin/invite-user/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete User",
                "description": "A user can be deleted by passing their user id.",
                "operationId": "inviteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id of the user whom the admin wants to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid user id"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user after being invited",
                "description": "",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Details of the user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Invalid token or invalid email address "
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Login Details of the user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Email address or password is invalid "
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs out current logged in user session",
                "description": "",
                "operationId": "logoutUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/user/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot user password",
                "description": "",
                "operationId": "forgotPassword",
                "requestBody": {
                    "description": "Email address of the user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/user/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset forgotten password",
                "description": "",
                "operationId": "resetPassword",
                "requestBody": {
                    "description": "Details of the user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "name",
                    "email",
                    "id",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Risabh Harry"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "test123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "test123"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "UserLoginBody": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "test123"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserLoginBody"
                }
            }
        }
    }
}